Moved image plotting functionality to TopicContainer
